rtkit-daemon: Restore error handling for DBus property Get
GCC caught this one:

rtkit-daemon.c: In function ‘dbus_handler’:
rtkit-daemon.c:1444:64: warning: comparison of constant ‘0’ with boolean expression is always false [-Wbool-compare]
                         if (!handle_dbus_prop_get(property, r) < 0) {
                                                                ^
rtkit-daemon.c:1444:64: warning: logical not is only applied to the left hand side of comparison [-Wlogical-not-parentheses]
rtkit-daemon.c:1444:29: note: add parentheses around left hand side expression to silence this warning
                         if (!handle_dbus_prop_get(property, r) < 0) {
                             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                             (                                 )

This caused reads for nonexistent properties to return nothing instead
of an error.

While we're at it, report the right error for this as well as bad
interfaces, too.