Make systemd integration (sd-daemon API calls) optional
Now that rtkit links against libsystemd in order to use the sd-daemon
API, it requires systemd to be installed - even if the system at
hand uses a different service manager and calling said API doesn't
actually do anything. Add a compile-time option ("enable-libsystemd" in
autoconf, "libsystemd" in meson) which controls whether to enable
systemd integration or not, in the latter case excluding all
sd-daemon-related bits from rtkit-daemon.

The logic of the new option is as follows:
 - by default, look for libsystemd and decide whether to enable this
   feature or not depending on whether the library has been found;
 - if explicitly disable, do not even bother looking for libsystemd;
 - if explicitly enabled, look for libsystemd and fail if it hasn't been
   found.

Note that this option only affects whether or not rtkit talks to systemd
via the sd-daemon API, i.e. it has no effect on the installation of unit
files.