Add logind system-suspend support
Add support for logind system suspend delay and signalling using DBus.

See https://www.freedesktop.org/wiki/Software/systemd/inhibit/

This adds a race-free way to suspend priorities when the system is
going to sleep, and resume them when the system wakes up. The main
reason for suspending priorities is the canary can be stopped, as a
system suspend-resume cycle frequently causes a false-positive of
the canary.